Adjusting variables a little
I was trying to put the variables in some kind of dictionary that could be processed easily by a backend to produce a table but I couldn't get that working.  I want to do a kind of background (maybe it will be based on the Ansible setup module). but I don't have time for that yet.