Creating new playbook
I wanted to introduce a variable that would control whether a failing command would abort a playbook.  I was changing a task that used to use `ignore_errors` but I switched to using `failed_when` because I couldn't get it to work with `ignore_errors`.  There seems to be an issue in Ansible - I found some tickets complaining about it but they didn't seem to want to address the issue.  I got it working with `failed_when` so I'm happy.